+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 - **BREAKING CHANGE**: `--verbose` (or `-V`) takes an int argument as log level (Verbosity level: 0 = silent, 1 = deletions only, 2 = detailed output)
 - **BREAKING CHANGE**: If no retention options are specified, an error is raise (before: `--last=10` was used as default)
 - Added --quartes option to keep the last N qarters
+- Added security checks
+- Fix some argument handling
 - Added CI workflow for linting and simple test
+- Added SECURITY.md and CONTRIBUTING.md
 
 ### 0.2.0 - 09.11.2025
 - **BREAKING CHANGE**: Added `--regex` option to explicit define type of pattern (default type of pattern is glob)